Don't add the "manual match number" to output
This leads to very noisy diffs when adding or removing manual matches in
the middle of the file, or when moving them around in the file.

We can now move the new NL entry to the position where it belongs.

The output sort order changed slightly, so the sums are a bit different
now (floating point addition is not associative).