Update old manual_matches with some insights from 2023 data
Turns out there is a separate ETS record for the gas-powered part of
Emile Huchet, so the "mixed fuels" issue can be solved. Also, data for
the Ineos / Inesco plant in Belgium and Amfard in France seems a bit
odd, so better ignore them. And Halmstadsverket might actually be the
"Gasturbiner", as the plant was renamed from "Gasturbiner Halmstad G12"
to "Halmstadverket G12" in ENTSO-E data in 2023.